removed warning in ExtensionRegistry
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5793 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/common/ExtensionRegistry.cpp
@@ -62,7 +62,7 @@ void ExtensionRegistry::registerExtensions(rw::common::Ptr<Plugin> plugin){
 		if(_descMap.find(desc.point)!=_descMap.end() ){
 			// check if plugin/extension allready exists
 			std::vector< std::pair<Extension::Descriptor, rw::common::Ptr<Plugin> > > &res = _descMap[desc.point];
-			for(int i=0;i<res.size();i++){
+			for(size_t i=0;i<res.size();i++){
 				if(res[i].first.id == desc.id && res[i].second->getId() == plugin->getId()){
 					continue;
 				}
